Use aqua GUI widgets
