--- qiskit_chemistry_ui/_emptyview.py
@@ -1,35 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2018 IBM.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# =============================================================================
-
-import tkinter as tk
-import tkinter.ttk as ttk
-
-
-class EmptyView(ttk.Frame):
-
-    def __init__(self, parent, **options):
-        super(EmptyView, self).__init__(parent, **options)
-        self._child = tk.Frame(self, background='white')
-        self._toolbar = ttk.Frame(self)
-
-    def grid(self, **options):
-        self._toolbar.pack(side=tk.BOTTOM, fill=tk.X)
-        self._child.pack(side=tk.LEFT, fill=tk.BOTH, expand=tk.TRUE)
-        ttk.Frame.grid(self, **options)
-
-    def set_toolbar_size(self, size):
-        self._toolbar.configure(width=size[0], height=size[1])
